Update readme and contribution guidelines
This new format is intended to tie the various aspects of the .NET SDK together into a cohesive whole. Additionally, these changes make the landing pages more visual and accessible, with a goal of better showcasing our excellent SDK resources.

Also, this change updates the contribution guidelines doc to match the formatting in other repos.